Restructured dashboard controller
