+++ dashboard/controller/dashboard/util/index.js
@@ -1,71 +1,6 @@
-// include libraries
 var request = require('request'),
-	moment = require('moment'),
-	common = require('../helper/common');
+	common = require('../../../helper/common');
 
-
-// init settings
-var ini = null;
-exports.init = function(settings) {
-	ini = settings;
-}
-
-// main landing page
-exports.index = function(req, res) {
-
-	// reinit l10n and moment with locale
-	if (req.query && req.query.lang) {
-		common.l10n.setLanguage(req.query.lang);
-		moment.locale(req.query.lang);
-	}
-
-	var data = {};
-
-	//get all users
-	exports.getAllUsers(req, res, function(usersResponse) {
-
-		// store
-		data.users = usersResponse;
-
-		//get all activities
-		exports.getAllActivities(req, res, function(activitiesResponse) {
-
-			// store
-			data.activities = activitiesResponse;
-
-			//get journal data
-			exports.getAllJournals(req, res, function(journalResponse) {
-
-				//process
-				var d = {}
-				d.totalJournal = journalResponse.length;
-				d.totalEntries = 0;
-				for (var i = 0; i < journalResponse.length; i++) {
-					d.totalEntries += journalResponse[i].count;
-				}
-
-				//store
-				data.journal = d;
-
-				//get all classrooms
-				exports.getAllClassrooms(req, res, function(classroomResponse) {
-					
-					//store
-					data.classrooms = classroomResponse;
-
-					// send to login page
-					res.render('dashboard', {
-						title: 'dashboard',
-						module: 'dashboard',
-						data: data,
-						account: req.session.user,
-						server: ini.information
-					});
-				});
-			});
-		});
-	});
-};
 
 exports.getAllJournals = function(req, res, callback) {
 	// call
@@ -85,8 +20,8 @@ exports.getAllJournals = function(req, res, callback) {
 				msg: common.l10n.get('ErrorCode'+body.code)
 			});
 		}
-	})
-}
+	});
+};
 
 exports.getAllActivities = function(req, res, callback) {
 	// call
@@ -106,8 +41,8 @@ exports.getAllActivities = function(req, res, callback) {
 				msg: common.l10n.get('ErrorCode'+body.code)
 			});
 		}
-	})
-}
+	});
+};
 
 exports.getAllUsers = function(req, res, callback) {
 	request({
@@ -131,7 +66,7 @@ exports.getAllUsers = function(req, res, callback) {
 			});
 		}
 	});
-}
+};
 
 exports.getAllClassrooms = function(req, res, callback) {
 	// call
@@ -152,4 +87,4 @@ exports.getAllClassrooms = function(req, res, callback) {
 			});
 		}
 	});
-}
+};
